requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.3
bitarray==3.3.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
cmdstanpy==1.2.5
contourpy==1.3.1
cycler==0.12.1
cytoolz==1.0.1
entrypoints==0.4
eth-abi==5.2.0
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
faker==37.1.0
favicon==0.7.0
fonttools==4.57.0
frozenlist==1.5.0
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
hexbytes==1.3.0
holidays==0.70
htbuilder==0.9.0
idna==3.10
importlib-resources==6.5.2
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.2
markdown==3.7
markdownlit==0.0.7
markupsafe==3.0.2
matplotlib==3.10.1
multidict==6.2.0
narwhals==1.34.1
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pillow==11.1.0
plotly==6.0.1
prometheus-client==0.21.1
propcache==0.3.1
prophet==1.1.6
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycryptodome==3.22.0
pydantic==2.11.3
pydantic-core==2.33.1
pydeck==0.9.1
pymdown-extensions==10.14.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyunormalize==16.0.0
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.1.0
rpds-py==0.24.0
rsa==4.9
six==1.17.0
smmap==5.0.2
soupsieve==2.6
st-annotated-text==4.0.2
st-theme==1.2.3
stanio==0.5.1
streamlit==1.44.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-shadcn-ui==0.1.18
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.1.2
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-extensions==4.13.1
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
web3==7.10.0
websockets==15.0.1
yarl==1.19.0
